+++ CHANGELOG.md
@@ -60,6 +60,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `object_info_cmd` for `compas_viewer.commends`.
 * Added `gridmode` to `GridObject`.
 * Added `checkbox` to `compas_viewer.components.SceneForm`.
+* Added `TextEdit` to handle `name` change.
+* Added `DefaultLayout` to handle gerneral `layout` setting to minimal.
+* Added `ColorButton` to manage and display a color of ViewerSceneObject.
+* Added `SettingLayout` to better manage complex layout with config input.
 
 ### Changed
 
@@ -78,6 +82,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed `compas_viewer.components.slider` step attribute.
 * Fixed `renderer.view` distorted init.
 * Fixed `tagobject` by adding back FreeSans.ttf.
+* Updated callback to `SceneTree`.
+* Updated `ObjectSetting` and `CameraSetting` to support setting from config.
 
 ### Removed